Add Automated Parameter Tuning support (#22)
* Improve CLI comments

* Fix error handling when an iteration fails.

* Improve color separation between methods. Improve error handling

* Add function to clean previous results for a test

* Fix -a flag in the clean command

* Fix X axis using Time instead of Value

* Fix maximum Y of CPU load plots

* Add more config options. Improve config show command.

* Change hard-coded parameters to read from the config file

* Update the drop and cut to account for delay

* Update Point-LIO to new rustle configuration

* Update A-LOAM to new rustle configuration

* Update Fast-LIO to new rustle format!

* Create transmission channel to pass messages between core and CLI. Add progress bars

* Update Faster-LIO to new rustle format!

* Fix confidence band error. Change from f64 to f32 to optimize the plotting algorithm.

* Handle most of Warnings.

* Add max and std ape,rpe to the detailed iterations table.

* Change the ros launch file to rustle-ros directory

* Add some unit tests

* Add Time efficiency metrics.

* Start implementing Robustness metrics

* Add duration time to dataset struct

* Before re-write the Test Definition

* Remove test definition service and db

* Add robustness metrics

* Fix minor issues

* Fix mismatch between drop/cut initial timestamps

* Update README, update project structure and clean redundant/outdated files

* Improve EVO error handling

* Fix Time Processing Frame formula

* Set the defaul value of EVO scale parameter to false.

* Fix test delete command not working

* Update example config for FAST-LIVO2

* Add KISS-ICP and FAST-LIVO2

* Add LIO-SAM speed, cut and drop

* Add text size option for bands. Increase default font sizes. Change default aspect ratio

* first commit by Francisco

* added function to change a single parameter of a SLAM config db record

* add tuning model

* Add Tune command

* Implement 'list' subcommand for 'tune' command

- Add TuningConfig(models), TuningRepo(db) and TuningService(services)

* Change message when list tuning configurations, when none are in the database

* Change table headers when listing tuning configurations

* Add parsing and validation code for tuning configurations

- in a tuning config file, parameters and their values can now be
specified inside the 'parameters' json object.
- in the 'parameters' object, for now, only values of type String, Bool
Number and Array are allowed. for json Objects(nested parameters),
simply pass the parameter, and the validation code will recursively
search for it.
- For arrays, i assume all elements to be of the same type.

* Add new errors, support for more tunable parameter types, and configurations

- More error types for clearer handling
- On grid search, configurations are now built

* Improve plotting logic

* Improve communication with ROS processes; Implement WebSockets

* Fix ground truth not being recorded after websockets update

* Create the groundtruth data when the dataset is added to RUSTLE; Read GT from rosbag directly

* Fix reading groundtruth from bag logic

* backup commit

* Bug fixing in random search, added iteration data csv export for simulated annealing

* More customization options for simulated annealing hyperparameters


---------

+++ core/src/models/messages.rs
@@ -3,7 +3,7 @@
 // I maybe need to add algorithm and test setup???
 #[derive(Debug, Clone)]
 pub struct ProgressMessage {
-    pub iteration_num: usize,
+    pub iteration_num: u64,
     pub algo: String,
     pub bag_time: f64,
     pub duration: f64,
